requirements.txt
altair
black
colorama
duckdb
flake8
ipykernel
isort
matplotlib
nox
numpy
openpyxl
pandas
plotly
polars
pyarrow
pylsp-mypy
python-dotenv
python-lsp-ruff
python-lsp-server
pytest
pytest-cov
pytest-mock
pytest-timeout
rich
scikit-learn
scipy
seaborn
sqlalchemy
statsmodels
xlsxwriter

requests
